backend/monitoring/advanced_monitoring.py
import os
import time
import logging
import platform
import psutil
import socket
from typing import Dict, Any, List
from datetime import datetime

import prometheus_client
from prometheus_client import Counter, Gauge, Histogram
import opentelemetry
from opentelemetry import trace
from opentelemetry.exporter.jaeger.thrift import JaegerExporter
from opentelemetry.sdk.resources import SERVICE_NAME, Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor
import threading

class AdvancedMonitoringSystem:
    """
    Comprehensive Monitoring and Observability System
    
    Features:
    - System Resource Monitoring
    - Performance Metrics
    - Distributed Tracing
    - Logging
    - Health Checks
    """
    
    def __init__(self, service_name: str = 'coinage'):
        """
        Initialize monitoring system
        
        Args:
            service_name: Name of the service being monitored
        """
        self.service_name = service_name
        
        # Prometheus Metrics
        self._setup_prometheus_metrics()
        
        # Distributed Tracing
        self._setup_distributed_tracing()
        
        # Logging Configuration
        self._configure_logging()
    
    def _setup_prometheus_metrics(self):
        """
        Set up Prometheus metrics for monitoring
        """
        # Request Metrics
        self.request_counter = Counter(
            'coinage_requests_total', 
            'Total number of requests',
            ['method', 'endpoint', 'status']
        )
        
        self.request_latency = Histogram(
            'coinage_request_latency_seconds', 
            'Request latency in seconds',
            ['method', 'endpoint']
        )
        
        # System Resource Metrics
        self.cpu_usage = Gauge(
            'coinage_cpu_usage_percent', 
            'CPU Usage Percentage'
        )
        
        self.memory_usage = Gauge(
            'coinage_memory_usage_bytes', 
            'Memory Usage in Bytes'
        )
        
        self.disk_usage = Gauge(
            'coinage_disk_usage_percent', 
            'Disk Usage Percentage'
        )
        
        # Trading Metrics
        self.trade_volume = Counter(
            'coinage_trade_volume_total', 
            'Total trading volume',
            ['asset_type']
        )
        
        self.trade_success_rate = Gauge(
            'coinage_trade_success_rate', 
            'Trading success rate'
        )
        
        # Network Metrics
        self.network_sent = Counter('coinage_network_bytes_sent', 'Network Bytes Sent')
        self.network_recv = Counter('coinage_network_bytes_recv', 'Network Bytes Received')
    
    def _setup_distributed_tracing(self):
        """
        Configure distributed tracing with OpenTelemetry and Jaeger
        """
        resource = Resource(attributes={
            SERVICE_NAME: self.service_name
        })
        
        trace.set_tracer_provider(TracerProvider(resource=resource))
        
        # Jaeger Exporter Configuration
        jaeger_host = os.getenv('JAEGER_HOST', 'localhost')
        jaeger_port = int(os.getenv('JAEGER_PORT', 6831))
        
        jaeger_exporter = JaegerExporter(
            agent_host_name=jaeger_host,
            agent_port=jaeger_port
        )
        
        trace.get_tracer_provider().add_span_processor(
            BatchSpanProcessor(jaeger_exporter)
        )
        
        self.tracer = trace.get_tracer(__name__)
    
    def _configure_logging(self):
        """
        Configure advanced logging
        """
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
            handlers=[
                logging.StreamHandler(),
                logging.FileHandler('coinage.log')
            ]
        )
    
    def record_request_metrics(
        self, 
        method: str, 
        endpoint: str, 
        status_code: int, 
        latency: float
    ):
        """
        Record request-related metrics
        
        Args:
            method: HTTP method
            endpoint: Request endpoint
            status_code: HTTP status code
            latency: Request latency
        """
        self.request_counter.labels(
            method=method, 
            endpoint=endpoint, 
            status=status_code
        ).inc()
        
        self.request_latency.labels(
            method=method, 
            endpoint=endpoint
        ).observe(latency)
    
    def record_system_metrics(self):
        """
        Collect and record system resource metrics
        """
        self.cpu_usage.set(psutil.cpu_percent())
        self.memory_usage.set(psutil.virtual_memory().used)
        self.disk_usage.set(psutil.disk_usage('/').percent)
        
        # Network Metrics
        network = psutil.net_io_counters()
        self.network_sent.inc(network.bytes_sent)
        self.network_recv.inc(network.bytes_recv)
    
    def record_trade_metrics(
        self, 
        asset_type: str, 
        trade_amount: float, 
        success: bool
    ):
        """
        Record trading-related metrics
        
        Args:
            asset_type: Type of traded asset
            trade_amount: Trade volume
            success: Trade success status
        """
        self.trade_volume.labels(asset_type=asset_type).inc(trade_amount)
        
        # Update trade success rate
        current_success_rate = self.trade_success_rate._value.get()
        new_success_rate = (current_success_rate + (1 if success else 0)) / 2
        self.trade_success_rate.set(new_success_rate)
    
    def start_trace(self, trace_name: str) -> Any:
        """
        Start a distributed trace
        
        Args:
            trace_name: Name of the trace
        
        Returns:
            Trace context
        """
        return self.tracer.start_as_current_span(trace_name)
    
    def get_system_health(self) -> Dict[str, Any]:
        """
        Retrieve comprehensive system health information
        
        Returns:
            System health details
        """
        return {
            'timestamp': datetime.utcnow().isoformat(),
            'hostname': socket.gethostname(),
            'platform': {
                'system': platform.system(),
                'release': platform.release(),
                'version': platform.version()
            },
            'cpu': {
                'usage_percent': psutil.cpu_percent(),
                'cores': psutil.cpu_count()
            },
            'memory': {
                'total': psutil.virtual_memory().total,
                'available': psutil.virtual_memory().available,
                'used_percent': psutil.virtual_memory().percent
            },
            'disk': {
                'total': psutil.disk_usage('/').total,
                'used': psutil.disk_usage('/').used,
                'used_percent': psutil.disk_usage('/').percent
            },
            'network': {
                'interfaces': list(psutil.net_if_stats().keys())
            }
        }
    
    def health_check(self) -> Dict[str, bool]:
        """
        Perform comprehensive health checks
        
        Returns:
            Health check results
        """
        checks = {
            'database_connection': self._check_database(),
            'redis_connection': self._check_redis(),
            'external_services': self._check_external_services()
        }
        
        return checks
    
    def _check_database(self) -> bool:
        """
        Check database connection
        
        Returns:
            Database connection status
        """
        try:
            # Implement database connection check
            return True
        except Exception:
            return False
    
    def _check_redis(self) -> bool:
        """
        Check Redis connection
        
        Returns:
            Redis connection status
        """
        try:
            # Implement Redis connection check
            return True
        except Exception:
            return False
    
    def _check_external_services(self) -> Dict[str, bool]:
        """
        Check external service connections
        
        Returns:
            External service connection statuses
        """
        external_services = {
            'market_data_api': self._check_market_data_api(),
            'payment_gateway': self._check_payment_gateway()
        }
        
        return external_services
    
    def _check_market_data_api(self) -> bool:
        """
        Check market data API connection
        
        Returns:
            Market data API connection status
        """
        try:
            # Implement market data API check
            return True
        except Exception:
            return False
    
    def _check_payment_gateway(self) -> bool:
        """
        Check payment gateway connection
        
        Returns:
            Payment gateway connection status
        """
        try:
            # Implement payment gateway check
            return True
        except Exception:
            return False

def start_prometheus_server(port: int = 8000):
    """
    Start Prometheus metrics server
    
    Args:
        port: Port to expose metrics
    """
    prometheus_client.start_http_server(port)

def main():
    """
    Demonstration of advanced monitoring
    """
    monitoring = AdvancedMonitoringSystem('coinage')
    
    # Start Prometheus metrics server
    start_prometheus_server()
    
    # Periodic system metrics collection
    def collect_system_metrics():
        while True:
            monitoring.record_system_metrics()
            
            # Get system health
            health = monitoring.get_system_health()
            logging.info(f"System Health: {health}")
            
            # Perform health checks
            health_checks = monitoring.health_check()
            logging.info(f"Health Checks: {health_checks}")
            
            time.sleep(60)  # Collect metrics every minute
    
    metrics_thread = threading.Thread(target=collect_system_metrics)
    metrics_thread.daemon = True
    metrics_thread.start()

    # Keep main thread running
    while True:
        time.sleep(1)

if __name__ == '__main__':
    main()
