backend/requirements.txt
# Core Flask Packages
flask-script==2.0.6
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-login==0.5.0
flask-bcrypt==1.0.1
flask-cors==3.0.10
python-dotenv==0.19.2

# Database
SQLAlchemy==1.4.29
redis==4.5.4

# Security
bcrypt==3.2.0
email-validator==1.1.3
pyjwt==2.6.0
pyotp==2.8.0

# Monitoring and Logging
prometheus-client==0.16.0
psutil==5.9.4
python-json-logger==2.0.7
elasticsearch==8.6.2

# Testing
pytest==6.2.5
coverage==6.2
hypothesis==6.75.0
faker==18.3.1

# Additional Utilities
requests==2.26.0
python-dateutil==2.8.2
pytz==2021.3
sentry-sdk==1.21.1

# Development and Deployment
gunicorn==20.1.0
werkzeug==2.2.3

# Performance and Caching
cachetools==5.3.0

# API and Validation
marshmallow==3.19.0
webargs==8.2.0

# Async and Concurrency
celery==5.2.7
gevent==22.10.2

# Monitoring and Tracing
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0

# Security Scanning
safety==2.3.4
bandit==1.7.5
